backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
antlr4-python3-runtime==4.9.2
